requirements.txt
flask==2.0.2
flask_inertia==0.3
flask_sqlalchemy==2.5.1
flask_migrate==3.1.0
flask_login==0.5.0
factory_boy==3.2.1
